+++ src/main/java/com/dinhkhang/code/service/IQRService.java
@@ -3,11 +3,13 @@ package com.dinhkhang.code.service;
 import com.dinhkhang.code.dto.QRSessionDTO;
 import com.dinhkhang.code.entity.QRSession;
 
+import java.math.BigDecimal;
+
 public interface IQRService {
-    QRSessionDTO generateQRSession(Long sessionId, Double teacherLat, Double teacherLong,
+    QRSessionDTO generateQRSession(Long sessionId, BigDecimal teacherLat, BigDecimal teacherLong,
             Integer expirationMinutes, Integer maxDistanceMeters);
 
     QRSession validateQRSession(Long qrId, String tokenSecret);
-    
+
     QRSession validateQRSessionBySessionIdAndToken(Long sessionId, String tokenSecret);
 }
