+++ src/main/java/com/dinhkhang/code/service/SecureAttendanceService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -52,9 +53,8 @@ public class SecureAttendanceService {
 
             if (!session.getStatus().equals(ClassSession.SessionStatus.IN_PROGRESS)) {
                 return AttendanceCheckInResponse.failed(
-                    "Buổi học chưa bắt đầu hoặc đã kết thúc",
-                    "FAILED_SESSION_NOT_ACTIVE"
-                );
+                        "Buổi học chưa bắt đầu hoặc đã kết thúc",
+                        "FAILED_SESSION_NOT_ACTIVE");
             }
 
             // CHECK 2: Token hợp lệ?
@@ -64,9 +64,8 @@ public class SecureAttendanceService {
 
             if (!qrSession.getClassSession().getId().equals(session.getId())) {
                 return AttendanceCheckInResponse.failed(
-                    "QR Code không khớp với buổi học",
-                    "FAILED_INVALID_QR"
-                );
+                        "QR Code không khớp với buổi học",
+                        "FAILED_INVALID_QR");
             }
 
             // CHECK 3: Email có trong lớp không?
@@ -81,45 +80,42 @@ public class SecureAttendanceService {
 
             if (!isInClass) {
                 return AttendanceCheckInResponse.failed(
-                    "Bạn không thuộc lớp học này. Vui lòng liên hệ giáo viên!",
-                    "FAILED_NOT_IN_CLASS"
-                );
+                        "Bạn không thuộc lớp học này. Vui lòng liên hệ giáo viên!",
+                        "FAILED_NOT_IN_CLASS");
             }
 
             // CHECK 4: Đã điểm danh chưa?
             if (attendanceRecordRepository.existsByStudentAndClassSession(student, session)) {
                 return AttendanceCheckInResponse.failed(
-                    "Bạn đã điểm danh cho buổi học này rồi",
-                    "FAILED_ALREADY_CHECKED"
-                );
+                        "Bạn đã điểm danh cho buổi học này rồi",
+                        "FAILED_ALREADY_CHECKED");
             }
 
             // CHECK 5: GPS trong bán kính?
             double distance = calculateDistance(
-                qrSession.getTeacherLatitude(),
-                qrSession.getTeacherLongitude(),
-                request.getLatitude(),
-                request.getLongitude()
-            );
+                    qrSession.getTeacherLatitude().doubleValue(),
+                    qrSession.getTeacherLongitude().doubleValue(),
+                    request.getLatitude().doubleValue(),
+                    request.getLongitude().doubleValue());
 
             if (distance > qrSession.getMaxDistanceMeters()) {
                 // Lưu record thất bại
                 saveFailedRecord(student, session, request, distance,
-                    AttendanceRecord.AttendanceStatus.FAILED_DISTANCE,
-                    String.format("Khoảng cách %.2fm vượt quá %.0fm", distance, (double)qrSession.getMaxDistanceMeters()));
+                        AttendanceRecord.AttendanceStatus.FAILED_DISTANCE,
+                        String.format("Khoảng cách %.2fm vượt quá %.0fm", distance,
+                                (double) qrSession.getMaxDistanceMeters()));
 
                 return AttendanceCheckInResponse.failed(
-                    String.format("Bạn ở quá xa (%.2fm). Giới hạn: %dm", distance, qrSession.getMaxDistanceMeters()),
-                    "FAILED_DISTANCE"
-                );
+                        String.format("Bạn ở quá xa (%.2fm). Giới hạn: %dm", distance,
+                                qrSession.getMaxDistanceMeters()),
+                        "FAILED_DISTANCE");
             }
 
             // CHECK 6: Device đã dùng chưa? (chống gian lận quét hộ)
             if (attendanceRecordRepository.findBySessionAndDeviceUid(session, request.getDeviceId()).isPresent()) {
                 return AttendanceCheckInResponse.failed(
-                    "Thiết bị này đã được sử dụng để điểm danh",
-                    "FAILED_DUPLICATE_DEVICE"
-                );
+                        "Thiết bị này đã được sử dụng để điểm danh",
+                        "FAILED_DUPLICATE_DEVICE");
             }
 
             // ✅ TẤT CẢ CHECK PASS - LƯU ĐIỂM DANH THÀNH CÔNG
@@ -128,7 +124,7 @@ public class SecureAttendanceService {
             record.setClassSession(session);
             record.setStudentLatitude(request.getLatitude());
             record.setStudentLongitude(request.getLongitude());
-            record.setDistanceMeters(distance);
+            record.setDistanceMeters(BigDecimal.valueOf(distance));
             record.setDeviceUid(request.getDeviceId());
             record.setFaceDataUrl(request.getSelfieBase64());
             record.setStatus(AttendanceRecord.AttendanceStatus.SUCCESS);
@@ -137,16 +133,14 @@ public class SecureAttendanceService {
             record = attendanceRecordRepository.save(record);
 
             return AttendanceCheckInResponse.success(
-                "✅ Điểm danh thành công! Khoảng cách: " + String.format("%.2fm", distance),
-                distance,
-                record.getId()
-            );
+                    "✅ Điểm danh thành công! Khoảng cách: " + String.format("%.2fm", distance),
+                    distance,
+                    record.getId());
 
         } catch (RuntimeException e) {
             return AttendanceCheckInResponse.failed(
-                "Lỗi: " + e.getMessage(),
-                "FAILED_INVALID_QR"
-            );
+                    "Lỗi: " + e.getMessage(),
+                    "FAILED_INVALID_QR");
         }
     }
 
@@ -154,16 +148,16 @@ public class SecureAttendanceService {
      * Lưu record thất bại để tracking
      */
     private void saveFailedRecord(User student, ClassSession session,
-                                   AttendanceCheckInRequest request,
-                                   double distance,
-                                   AttendanceRecord.AttendanceStatus status,
-                                   String reason) {
+            AttendanceCheckInRequest request,
+            double distance,
+            AttendanceRecord.AttendanceStatus status,
+            String reason) {
         AttendanceRecord record = new AttendanceRecord();
         record.setStudent(student);
         record.setClassSession(session);
         record.setStudentLatitude(request.getLatitude());
         record.setStudentLongitude(request.getLongitude());
-        record.setDistanceMeters(distance);
+        record.setDistanceMeters(BigDecimal.valueOf(distance));
         record.setDeviceUid(request.getDeviceId());
         record.setFaceDataUrl(request.getSelfieBase64());
         record.setStatus(status);
@@ -189,4 +183,3 @@ public class SecureAttendanceService {
         return EARTH_RADIUS * c;
     }
 }
-
